Auto-kink trajectories
Shooting darts and stuff happens in a straight line as always, but now
it can aim for a point that is not the exact center of the target cell,
if that will avoid hitting a wall or an obstacle along the way.

This makes it unnecessary to aim for a cell beyond your intended
target, trying to trick the Bresenham algorithm into passing through
the cell you want.